Update to Flixel v2.55
Previous version was Flixel v2.35.

Also moved the library from the `src	` folder to an easy to use SWC.

-------

The `revenge.asproj` file (the project file used by FlashDevelop) may
needed to be edited to reflect the addition of the Flixel SWC.

Since I am in Linux, I am unable to edit the project file myself.